move playlist out of presentational
