+++ client/src/App.tsx
@@ -5,7 +5,7 @@ import Home from './components/Home';
 import Landing from './components/Landing';
 import styled from '@emotion/styled';
 import { AppBar, Typography, Toolbar } from '@mui/material';
-import Playlist from './components/presentational/Playlist';
+import Playlist from './components/Playlist';
 
 function App() {
